# SkillFactory_Comments/internal/storage/queries/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package queries

import (
	"time"

	"github.com/jackc/pgx/v5/pgtype"
)

type Comment struct {
	ID              int32
	NewsID          pgtype.Int4
	ParentCommentID pgtype.Int4
	Content         string
	CreatedAt       time.Time
}